Last safe commit before last minute big massive migration to a working state. This will not be fun.

+++ Bytes/AKData.h
@@ -10,7 +10,7 @@
 
 @interface AKData : NSObject
 
-+(int)createCodeWithBytes:(int)bytes;
-+(int)getBytesWithCode:(int)bytes;
++(int)createCodeWithBytes:(NSInteger)bytes;
++(int)getBytesWithCode:(NSInteger)bytes;
 
 @end
